+++ src/aux-auth/shared/AuthMetadata.ts
@@ -1,6 +1,7 @@
 export interface UserMetadata {
     email: string;
     name: string;
+    phone: string;
     avatarUrl: string;
     avatarPortraitUrl: string;
 }