+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # CasualOS Changelog
 
+## V3.0.2
+
+#### Date: 3/16/2022
+
+### :boom: Breaking Changes
+
+-   Removed the following functions:
+    -   `server.exportGpio()`
+    -   `server.unexportGpio()`
+    -   `server.getGpio()`
+    -   `server.setGpio()`
+
+### :rocket: Improvements
+
+-   Improved performance for lower end devices by making CasualOS more efficient when automatically updating bots with user input.
+-   Added the ability to login with a phone number instead of an email address.
+    -   This feature is enabled by the `ENABLE_SMS_AUTHENTICATION` environment variable during builds.
+-   Added the ability to automatically synchronize device clocks and expose the synchronized information to scripts.
+    -   The following properties have been added:
+        -   `os.localTime` - The local clock time in miliseconds since the Unix Epoch.
+        -   `os.agreedUponTime` - The synchronized clock time in miliseconds since the Unix Epoch.
+        -   `os.instLatency` - The average latency between this device and the inst in miliseconds. Smaller values are generally better.
+        -   `os.instTimeOffset` - The delta between the local time and agreed upon time in miliseconds.
+        -   `os.instTimeOffsetSpread` - The uncertainty of the accuracy of the `os.instTimeOffset` value. Measured in miliseconds. Smaller values indicate that `os.agreedUponTime` is more accurate, larger values indicate that `os.agreedUponTime` is less accurate.
+        -   `os.deadReckoningTime` - The synchronized clock time that includes an additional 50ms offset to try to ensure that all devices are synchronized once the time ocurrs.
+-   Improved `animateTag()` to support custom easing functions and a custom start time.
+    -   The `easing` property in the options object that is passed to `animateTag()` now supports custom functions for custom easing behaviors. The function should accept one parameter which is a number between 0 and 1 that represents the progress of the animation and it should return a number which is the value that should be multiplied against the target tag. See the documentation of `animateTag()` for an example.
+    -   The `startTime` property is now supported in the options object that is passed to `animateTag()`. It should be the number of miliseconds since the Unix Epoch that the animation should start at. For example, `os.localTime + 1000` will cause the animation to start in 1 second.
+
+### :bug: Bug Fixes
+
+-   Fixed an issue where `bot.vars` would get cleared after the scripts that created it finished their initial execution.
+-   Fixed an issue where `labelColor` did not work on menu bots that had `form` set to `input`.
+-   Fixed an issue where `labelColor` would not work unless the menu bot had a `label`.
+    -   This is useful for menu bots that only use icons.
+
 ## V3.0.1
 
 #### Date: 2/17/2022